refactor: move remote mods into different folders
